First unit test
